+++ src/components/Dashboard/SideBar/Navigation.jsx
@@ -1,24 +1,117 @@
+import { NavLink } from 'react-router-dom';
+import {ReactComponent as NavCurrency} from "../../../images/icons/NavCurrency.svg";
+import { ReactComponent as NavDiagram } from "../../../images/icons/NavDiagram.svg";
+import { ReactComponent as NavHome } from "../../../images/icons/NavHome.svg";
 import styled from "styled-components";
-import { NavLink } from "react-router-dom";
 
-const RouteButtons = styled.div`
-  width: 100px;
-  height: 75px;
-`;
-const RouteButton = styled.button`
-  font-size: 15px;
-  margin-bottom: 15px;
+const HomeSvg = styled(NavHome)`
+    width: 38px;
+    height: 38px;
+    & path {
+      fill: #6E78E8;
+    }
+
+    margin-right: 36px;
+
+    @media screen and (min-width: 480px) {
+    width: 18px;
+    height: 18px;
+    margin-right: 23px;
+  }
 `;
 
-export const Navigation = () => {
+const HomeLink = styled.div`
+    display: flex;
+    align-items: center;
+    padding-bottom: 12px;
+`;
+
+const PageText = styled.span`
+display: none;
+font-family: 'Poppins';
+font-style: normal;
+font-weight: 400;
+font-size: 18px;
+line-height: 1,5;
+color: #000000;
+
+
+@media screen and (min-width: 480px) {
+    display: block;
+  }
+`;
+
+const DiagramSvg = styled(NavDiagram)`
+    width: 38px;
+    height: 38px;
+    & path {
+      fill: #6E78E8;
+    }
+
+    margin-right: 36px;
+
+    @media screen and (min-width: 480px) {
+    width: 18px;
+    height: 18px;
+    margin-right: 23px;
+`;
+
+const DiagramLink = styled.div`
+    display: flex;
+    align-items: center;
+`;
+
+const CurrencySvg = styled(NavCurrency)`
+    width: 38px;
+    height: 38px;
+    & path {
+      fill: #6E78E8;
+    }
+
+  @media screen and (min-width: 480px) {
+    display: none;
+  }
+`;
+
+const NavList = styled.div`
+    display: flex;
+    justify-content: center;
+    padding-top: 15px;
+    padding-bottom: 31px;
+
+    @media screen and (min-width: 480px) {
+    flex-direction: column;
+
+    padding-top: 40px;
+    padding-bottom: 28px;
+  }
+`;
+
+export default function Navigation() {
   return (
-    <RouteButtons>
-      <RouteButton>
-        <NavLink to="home">Home</NavLink>
-      </RouteButton>
-      <RouteButton>
-        <NavLink to="diagram">Statistic</NavLink>
-      </RouteButton>
-    </RouteButtons>
+    <>
+    <NavList>
+        <NavLink
+          to="home"
+        >
+          <HomeLink>
+            <HomeSvg src={NavHome} />
+            <PageText>Главная</PageText>
+          </HomeLink>
+      </NavLink>
+          <NavLink
+            to="diagram"
+        >
+          <DiagramLink>
+          <DiagramSvg src={NavDiagram} />
+            <PageText>Статистика</PageText>
+            </DiagramLink>
+          </NavLink>
+          <NavLink
+            to="currency"
+      ><CurrencySvg src={NavCurrency}/>
+      </NavLink>
+      </NavList>
+    </>
   );
-};
+}
