--- src/components/Diagram.jsx
@@ -1,10 +1,0 @@
-import { Currency } from "../components/Currency";
-
-export default function Home() {
-  return (
-    <>
-    <h1>Diagram</h1>
-      <Currency />
-    </>
-  );
-}
