--- src/components/Home.jsx
@@ -1,10 +1,0 @@
-import { Currency } from "../components/Currency";
-
-export default function Home() {
-  return (
-    <>
-    <h1>Home</h1>
-      <Currency />
-    </>
-  );
-}
